Refactor parameter mappers
