+++ ballerina-to-openapi/src/main/java/io/ballerina/openapi/service/mapper/parameter/PathParameterMapper.java
@@ -19,15 +19,15 @@ package io.ballerina.openapi.service.mapper.parameter;
 import io.ballerina.compiler.api.SemanticModel;
 import io.ballerina.compiler.api.symbols.PathParameterSymbol;
 import io.ballerina.compiler.api.symbols.TypeSymbol;
-import io.ballerina.compiler.syntax.tree.Node;
-import io.ballerina.compiler.syntax.tree.SyntaxKind;
+import io.ballerina.compiler.api.symbols.resourcepath.util.PathSegment;
 import io.ballerina.openapi.service.mapper.AdditionalData;
 import io.ballerina.openapi.service.mapper.diagnostic.DiagnosticMessages;
 import io.ballerina.openapi.service.mapper.diagnostic.IncompatibleResourceDiagnostic;
 import io.ballerina.openapi.service.mapper.diagnostic.OpenAPIMapperDiagnostic;
+import io.ballerina.openapi.service.mapper.model.OperationAdaptor;
 import io.ballerina.openapi.service.mapper.type.TypeMapper;
+import io.ballerina.tools.diagnostics.Location;
 import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.parameters.Parameter;
 import io.swagger.v3.oas.models.parameters.PathParameter;
 
 import java.util.List;
@@ -36,7 +36,7 @@ import java.util.Map;
 import static io.ballerina.openapi.service.mapper.utils.MapperCommonUtils.removeStartingSingleQuote;
 import static io.ballerina.openapi.service.mapper.utils.MapperCommonUtils.unescapeIdentifier;
 
-public class PathParameterMapper implements ParameterMapper {
+public class PathParameterMapper extends AbstractParameterMapper {
 
     private final TypeSymbol type;
     private final String name;
@@ -44,9 +44,15 @@ public class PathParameterMapper implements ParameterMapper {
     private final SemanticModel semanticModel;
     private final List<OpenAPIMapperDiagnostic> diagnostics;
     private final OpenAPI openAPI;
+    private final Location location;
+    private final PathSegment.Kind pathSegmentKind;
 
     public PathParameterMapper(PathParameterSymbol pathParameterSymbol, OpenAPI openAPI, Map<String, String> apiDocs,
-                               SemanticModel semanticModel, List<OpenAPIMapperDiagnostic> diagnostics) {
+                               OperationAdaptor operationAdaptor, SemanticModel semanticModel,
+                               List<OpenAPIMapperDiagnostic> diagnostics) {
+        super(operationAdaptor);
+        this.location = pathParameterSymbol.getLocation().orElse(null);
+        this.pathSegmentKind = pathParameterSymbol.pathSegmentKind();
         this.type = pathParameterSymbol.typeDescriptor();
         this.openAPI = openAPI;
         this.name = unescapeIdentifier(pathParameterSymbol.getName().get());
@@ -55,22 +61,14 @@ public class PathParameterMapper implements ParameterMapper {
         this.diagnostics = diagnostics;
     }
 
-    /**
-     * Map path parameter data to OAS path parameter.
-     */
-    public void mapPathParameter(List<Parameter> parameters, Node pathParam) {
-        if (!pathParam.kind().equals(SyntaxKind.RESOURCE_PATH_REST_PARAM)) {
-            parameters.add(getParameterSchema());
-        } else {
-            DiagnosticMessages errorMessage = DiagnosticMessages.OAS_CONVERTOR_125;
-            IncompatibleResourceDiagnostic error = new IncompatibleResourceDiagnostic(errorMessage,
-                    pathParam.location(), pathParam.toString());
-            diagnostics.add(error);
-        }
-    }
-
     @Override
     public PathParameter getParameterSchema() {
+        if (pathSegmentKind.equals(PathSegment.Kind.PATH_REST_PARAMETER)) {
+            DiagnosticMessages errorMessage = DiagnosticMessages.OAS_CONVERTOR_125;
+            IncompatibleResourceDiagnostic error = new IncompatibleResourceDiagnostic(errorMessage, location, name);
+            diagnostics.add(error);
+            return null;
+        }
         PathParameter pathParameter = new PathParameter();
         pathParameter.setName(name);
         pathParameter.setRequired(true);
