+++ ballerina-to-openapi/src/main/java/module-info.java
@@ -40,5 +40,6 @@ module io.ballerina.openapi.service {
     exports io.ballerina.openapi.service.mapper.utils;
     exports io.ballerina.openapi.service.mapper.parameter.model;
     exports io.ballerina.openapi.service.mapper.constraint.model;
+    exports io.ballerina.openapi.service.mapper.parameter;
 }
